Let the user confirm the deletion before taking it to action

+++ client/app/js/board_list.jsx
@@ -47,7 +47,9 @@ export default class BoardList extends React.Component {
 
   onIconClickHandler = (event, id) => {
     event.stopPropagation();
-    BoardActions.deleteBoard(id);
+    BoardActions.showConfirmation(() => {
+      BoardActions.deleteBoard(id);
+    });
   }
 
   render() {
